requirements/base.txt
Django>=4.2
djangorestframework
psycopg2-binary

celery
redis
django-redis
requests

drf-yasg

scikit-learn
lightgbm
pandas
numpy
imblearn

python-dotenv

prometheus-client
django-prometheus

ollama